+++ lib/sparql/extension_function/registration.ex
@@ -1,0 +1,5 @@
+import ProtocolEx
+
+defprotocol_ex SPARQL.ExtensionFunction.Registration do
+  def extension_function(name), do: nil
+end