+++ lib/sparql/extension_function/registry.ex
@@ -3,53 +3,18 @@ defmodule SPARQL.ExtensionFunction.Registry do
   A registry of all available `SPARQL.ExtensionFunction`s.
   """
 
-  @key :sparql_extension_functions
-
-  def init do
-    FastGlobal.put(
-      @key,
-      find_available_extension_functions!()
-      |> Map.new(fn extension -> {extension.name, extension} end)
-    )
-  end
+  alias SPARQL.ExtensionFunction.Registration
 
   @doc """
   Returns the extension for the given name.
 
   If no function for the given name exists, `nil` is returned.
   """
-  def get_extension(name)
+  def extension_function(name)
 
-  def get_extension(%RDF.IRI{value: value}), do: value |> get_extension()
+  def extension_function(%RDF.IRI{value: value}), do: value |> extension_function()
 
-  def get_extension(name) do
-    FastGlobal.get(@key)[name]
-  end
-
-  @doc """
-  Returns a map of all extensions by name.
-  """
-  def get_all() do
-    FastGlobal.get(@key)
-  end
-
-  defp find_available_extension_functions!() do
-    SPARQL.ExtensionFunction
-    |> SPARQL.Utils.find_behaviour_implementations()
-    |> check_homonyms!()
-  end
-
-  defp check_homonyms!(extensions) do
-    extensions
-    |> Enum.group_by(fn extension -> extension.name end)
-    |> Enum.reject(fn {_, extension} -> Enum.count(extension) == 1 end)
-    |> case do
-         [] -> extensions
-         homonyms ->
-           raise """
-            The following names are used for multiple SPARQL.ExtensionFunctions:
-            #{inspect Map.new(homonyms)}
-            """
-       end
+  def extension_function(name) do
+    Registration.extension_function(name)
   end
 end